// apps/notification-service/src/interfaces/iLog.interface.ts
import { JsonObject } from "../dto/outboxEvents/createOutboxEvent.dto.ts";

export interface ILogMetadata {
    service: string;
    event?: string;
    event_id?: string;
    entity_id?: number;
    method?: string;
    payload?: JsonObject;
    created_at: string;
    error_name?: string;
    error_message?: string;
    error_stack?: string;
}